Implement partial r-indicators and coefficients of variation
Implements unconditional and conditional partial r-indicators and
coefficients of variation, by variable and by category. Also changes the
`job` variable in the HLC data from logical to factor, because partial
r-indicator and coefficient of variation calculations require selected
variables to be categorical.

The implementation is tested against RISQ 2.1 reference implementation.