Make conditional partials 0 for variables outside model
The conditional partial representativity indicator and coefficient of
variation should be 0 for variables that are not part of the response
model. The calculation of these values could lead to unintended results
due to divisions with very small numbers at machine precision.

In this change, we set for conditional partials to 0 for variables
outside the model. Unconditional partials are not affected.