Make conditional partials NA for lone variable in model
The conditional partial representativity indicator and coefficient of
variation require other variables to condition on. Therefore, these
values cannot be calculated for a variable if that variable is the only
variable in the predictor of the model.

In this change, we set conditional partials to NA when there are no
other variables to condition on. Unconditional partials are unaffected.